client/requirements.txt
requests==2.25.1
websocket-client==1.0.1
python-socketio==5.3.0
vgamepad==0.0.2